change fastboot policies based on upstream feedback
Fastboot is tightly integrated with kernelflinger. It is only
an implementation detail that they are not the same binary.
We need to treat it as a special case and not try to integrate
loading it with the rest of the verified boot policy.

- Fastboot no longer lives in a partition as raw data, it is
now a file on the EFI system partition. Flashing the bootloader
updates both the loader and Fastboot at the same time.

- Fastboot image is always verified with the OEM keystore embedded
in the loader, regardless of device or boot state.

- Some device state UX needs to go into Fastboot mode instead of
recovery, as that is the only mechanism to clear/update keystores
or change the device state; Recovery can't do this.

- For bootable USB media, we now have a sentinel file which will
instruct the loader to always enter Fastboot mode.

Change-Id: If95aff98c01798dd87c4b3269a0376ba11af8c0a
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: 223084