main() cleanup, add insecure compilation, expand some comments
- main() recieved some cleanup, in particular the UX for EFI
secure boot being disabled now happens after magic key detection
and not before it.

- inevitably BSP developers are going to need a loader which doesn't
ruin their day with non-cacheable UX prompts. If -DINSECURE is in
CFLAGS when kernelflinger is built, all UX is skipped and the device
will always boot into RED state, complaining to the console as it
does so. It's intended that this insecure variant will be populated
only in 'eng' builds.

- Clarified some UX that the boot image may be corrupted instead of
unverified (can happen if loader is unlocked)

- More comments added to assist with review.

Change-Id: I6357fe97ac84e3ffa9738852d38a8786720b146d
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>