Replace uefi_shim by libopenssl
libopenssl relies on some symbols to be defined by external libraries.
To be able to link properly we have to define all of those symbols, by
stubbing the one which are useless in our EFI case and implementing the
needed ones.

Since libkernelflinger was relying on uefi_shim library, we had also to
move some basic functions definitions such as string manipulations into
the libkernelflinger itself.

Change-Id: I6dd8dbdbef976f24401d636e4fd082f073de76ab
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-17735
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>
Reviewed-on: https://android.intel.com:443/473644