remove libkernelflinger dependencies on libfastboot
We don't want these libs to be mutually dependent, as it simplifies
the build to omit libfastboot on devices that use Userfastboot.

The functions moved relate to state variables that are used regardless
of whether EFI fastboot, or Userfastboot, is in use. Move these to
libkernelflinger so they are universally available.

Change-Id: I161c84ee214bc0fab38690ac5639fc89edfaf015
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>