implement 'fastboot continue' as a reboot
Currently when returning out of a Fastboot session there is some
logic to continue the boot, but it doesn't handle a lot of edge
cases such as if a UX needs to be displayed, or the selected
keystore was changed.

BootloaderRequirements 1.8 has no mention of 'fastboot continue'
at all, and we found on Nexus device it just resets. Do that here
too, so that we guarantee the correct boot logic.

Change-Id: Ic0a7cfe00d060590bdd4481d1408bfa4eca3d3c6
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>