initialize display on demand; refactor APIs
ui_init() causes the screen to be cleared, which is undesirable
on devices that have a splash screen drawn by the BIOS. UX code
is changed to only call ui_init() on demand. This function is
idempotent so we don't need to track initialization.

Combine some of the logic in ui_default_screen() into ui_init(),
and split out the functionality to display a vendor splash screen
into its own function. The policy is to display a splash on bootup
if a particular EFI variable is set, and also display it when
exiting from a Fastboot session.

Change-Id: If2247a5e04d6a70a5fbc8d773e1fff2e3d5e1f55
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>