add 'fastboot oem reprovision'
Align with Userfastboot 08.00; this command is only present on non-user
builds and will put the device back into provisioning mode and reboot
to dnx.

Allows a User build to be provisioned on a device that already has
software on it without being hassled by Factory Unlock Protection;
the procedure is to boot a non-user loader, run "fastboot oem reprovision",
and then provision the User build with PhoneFlashTool.

Change-Id: I274ec43f9f1b569c0d3076524511e6090e5ee457
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>