simplify magic key detection if using EFI Fastboot
If we are using EFI Fastboot, magic key will unconditonally
boot into Fastboot mode. If the user wants to enter the
Recovery Console they can do so from the UI.

The code which enforces that the key is released has been
moved to ui_wait_for_input() as it is something we would
want for any use of that function. The boot is now not blocked
on key release unless a UX is shown.

Change-Id: I67dfa9924a8bfca2a58f5f66c6092815a9f95b6a
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>