add an error code for each UX prompt
In practice, when users are having a problem which results in a
loader UX being displayed, they almost never report what specific
message is presented, they just complain "I have to hit Volume Up".

Add specific error codes to the UX so they it's easy for users
to communicate the specific problem they are having.

Change-Id: I38a9f0ca674d2bbac33ed7db56c27c523bf062d1
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>