Fix validate_bootimage in case of ota multistage
OTA multistage need to flash the new recovery image on the boot
partition and reboot on it before continuing the update. In this
case, the boot_target is NORMAL_BOOT but the tag inside the image
is recovery and not boot.
To validate the NORMAL_BOOT, we allow both boot and recovery tag.

Change-Id: I769e0695a80d125150b208a5cba6165fa4af23af
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-11223
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>