boot flow: check battery level
The platform boots in CHARGING mode when the battery level is in deep
discharge state (bellow the threshold defined in the OEM1 ACPI table)
and a charger is plugged-in.  If the battery is in deep discharge
state and no charger is plugged-in, the device turns off.

This patch introduces the Google BootloaderRequirements 2.0
battery-level fastboot variable.

Change-Id: Ia4a677429d8fdd3d75bc4b9bc61a10bd29267409
Tracked-On: https://jira01.devtools.intel.com/browse/AREQ-3333
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>