Fix fastboot_info() with usb_write()
-> usb_write() must be called only once, and then
we must return the callback function and let usb driver do it's stuff.

fastboot_info() is now buffered to the end of the command.
each info is then sent one by one.

introducing a new state "STATE_TX" which means there are data on
the buffer that should be sent to host.

remove getvar worker (it can now use the buffered mechanism)

Change-Id: Iaf44c2e3aa8d0bae59a3d98a5dde5a09d9b5b8a6
Signed-off-by: Jocelyn Falempe <jocelyn.falempe@intel.com>