flash: buffer hunks
Android sparse files are usually constituted of a lot of small hunks.
A quick analysis of a 2GB system image shows that :
 - 34% of the hunks are 4096 bytes
 - 97% of the hunks are less than 1MB
Small hunks are now accumulated into a buffer.  Once the buffer is
full, it is flashed.

Measurements show that flashing an ordinary 2GB system image is 10%
faster with this algorithm.

Change-Id: I8c33dff06e5b78bb7acbb55e472aeb4ecf7c3ef2
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>