Change confirmation UX
Add the possibility of handling UI with 3 buttons instead of 2
by adding the support of power button, and use it to offer a
yes/no menu for UX confirmation.
The highlight functionality is added so that vol.up and down
are used to move the selection and power button to select it.

Change-Id: I9780e3a2765f716d9c2ed39a014659c14fdc78be
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-5429
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>