fastboot: fix size on string descriptor zero
This descriptor needs to be 4 bytes long if one language is
supported, BIOS interprets longer sizes as multiple languages
supported.

sizeof(LANG_EN_US) is evaluated to be an unsigned int (4 bytes),
because of:

	2 + sizeof(LANG_EN_US)

...kernelflinger is passing 0x6 as the size, which ends up duplicating
language entries. This broke the Manufacturing Flash Tool logic
to always read the 4th string as the SerialNumber as all the strings
were duplicated.

To fix the implementation, LANG_EN_US is casted to a 2 byte value
which is the expected behavior.

Change-Id: I06b619904b462c6629274973feaa4212ef538244
Signed-off-by: Omar Ramirez Luna <omar.r.ramirez.luna@intel.com>
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>