Add UFS support
As Broxton comes with UFS as mass storage in addition of eMMC,
kernelflinger must be able to support it.

This patch adds support of the SCSI IO protocol to be able to erase
UFS blocks using UNMAP command, which is equivalent to eMMC secure
erase.

This patch also abstracts the eMMC GPP partitions into logical unit
which can be either a GPP for an eMMC, or a LUN for a UFS.
Once for all, the implementation to be used is determined at the first
storage operation.

Change-Id: I2e0b8eb5a407a8b36d40152e3acea7806d7266cc
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-6963
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>