add "no oem unlock" build-time policy
If this flag is set, the following policy changes are in effect:

- The device will refuse to boot if the keystore is modified, UEFI
  secure boot is turned off, or Fastboot is somehow in an unlocked
  state outside of provisioning mode
- The UXes have been adjusted to reflect the above policy
- "fastboot oem unlock" always fails if the device isn't in provisioning
  mode and is a User build.

The device shouldn't boot to the Android session if a lock state hasn't
been set. This is necessary for device policies where unlocking the device
is forbidden in case the OEM unlock state variables have been tampered with.
We only enforce this in user builds, as some development
scenarios involve dediprog-ing a new BIOS which will reset these state
variables.

Future development will alter this to a runtime policy which can be modified
with authenticated tokens.

The only way to unlock a device with a User build is with a dediprog.

Change-Id: If54ffd4a1f474a436e49fda411f9697ecc67ecb7
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>