allow 'oem setvar' debug override of kernel command line
If the user sets AppendCmdline, PrependCmdline, or ReplaceCmdline,
the boot image command line will persistently be appended, prepended,
or replaced with the supplied value. These options may be used in
any combination.

This only takes effect for "normal" boots, and a warning is
printed if any modification is done. This only works in non-user
builds.

Issue: GMINL-5552
Change-Id: Ia6a9f148a899c9f367c36da4e97a65f13e05c1c4
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>