adjust UX policy to M-dessert specification
The online specification for Verified Boot on source.android.com
has relaxed the UX requirements and now all UXes, regardless of
the detected issue, work as follows:

- If no user input, boot proceeds after a 5 second timeout
- If the user presses Volume Up, boot proceeds immediately
- If the user presses any other detectable key, the timeout
  is canceled and the message remains until Volume Up is
  pressed

There is no longer a choice of alternate boot targets or fallback
to Recovery, so the logic in efi_main() needed to be simplified.

"user is the enemy" devices that have NO_DEVICE_UNLOCK enabled
power off instead of proceeding.

Issue: GMINL-8401
Change-Id: I2946a9816b0c18f22b6d758eb1f87f5dc5aa5923
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>