fastboot: Dual storage support
With Broxton, two PCI storage are available: UFS and eMMC.
When coming from DnX we are not able to know on which one we have to do
flashing operations.
To align on PC way of doing things, the user have to determine on which
storage he wants to install the system.
For that, we introduce a fastboot oem set-storage command which will
select the primary storage to use for the following flashing commands.

Change-Id: I85d55d3b63530d50519f8f75ddc40abacecc15bf
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>