fastboot: fix state-machine race condition
We have changed the BIOS stack design in order to handle the USB
SETUP (EP0) packets while we are calling time consuming block IO
functions.  As a consequence, kernelflinger is no longer
mono-threaded.  The way we handle the event makes us safe (no
read/write concurrency).  But the state machine behavior is impacted
and the current state can be changed before we exit the current
command processing.

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-33199
Change-Id: I7b5d6a5ea69825d3e5f41dc20731f1029294bb92
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>