set_efi_variable: delete variable only for different attributes
The current set_efi_variable implementation always delete the variable
before setting the new value despite the fact that the variable might
have the same attributes.  If the new variable creation fails we end
up with no variable at all which is especially damageable for the
OEMLock variable.  For instance a device transition from lock state to
verified state can end up into a provisioning mode which shall never
been reachable with a USER build.

We no more delete the variable if the attributes are the same.

Tracked-On: IMINAN-36396
Change-Id: Idca30040ac802219a5ff4964077e1a845912a44a
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>