blobstore: add blobstore_get_addr() function
We already have all the blobstore data in RAM so having blobstore_get()
force us to copy all the data on the heap is a little wasteful. Add
a function which just returns the address and size of the data we
want, so that it doesn't need to be freed later.

Change-Id: Iaa777b2627a0e20ebe4818d3ba19c92299708383
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>