set kernel command line variables from blobstore
* A new blobstore type BLOB_TYPE_BOOTVARS is introduced.
It's a text file, each line is inserted into the kernel command line.
This allows the kernel command line to be customized on a per
board basis using the same boot image. The first use-case is to
insert autodetect-related information for SoFIA targets.

* text parser now supports a context pointer so that callback functions
don't need to use globals

* we now have a generic get_bootimage_blob() function which returns
a pointer to blob data (for the specified type)  within the bootimage data.
It doesn't need to be freed, it's within the existing bootimage and
not a copy.

* BLOBSTORE_BLOB_NOT_FOUND errors reduced to debug statements

Change-Id: I7756473e2ce75b06ee06e8add4bda2e85fd6abc7
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>