blobstore: rewrite
The existing blobstore implementation had some fundamental problems
which were simpler to remedy with a redesign:

- The sizes of the metablocks were dependent on the number of different
  blob types since the last element of the enumerated type determined
  the size of an embedded array. If support for new types were added,
  older loaders would see a corrupted blobstore.
- Very little was pre-calculated in the blobstore data. Upon reading, not
  only did all the metablocks need to be recreated on the heap, but the
  hashes for all the items in the blobstore had to be computed at runtime.
- The codebase was in excess of 2000 lines of code, which made it annoying
  to port to other loaders. There were a lot of layers of abstraction which
  weren't necessary given the intended use, and we didn't need any code
  to modify the blobstore at runtime.

What we really wanted was something simple, where everything was pre-calculated
at build time, and access to the blobstore could be done very quickly without
building new data structures on the heap.

The new, simpler design has an array-based hash table embedded directly
in the blobstore data and takes into account both the board name and type
when computing the hash. We just need to pull data out of a blobstore already
loaded into RAM, so the code is now just 150 lines.

More detail on the layout of this new data structure can be found in
the docstring for device/intel/build/blobstore.py

The existing unit test is not applicable since blobstores can't be created
at runtime, only read.

Change-Id: Ia7f6e37de74fc80b732c2f588c8e6ea77dc5f3a4
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>