optimize setting OEM vars
Setting nonvolatile EFI variables is *slow*. We really don't need this to
happen at every boot. Instead, base it on whether an UPDATE_OEM_VARS
variable is set, and if so, clear it and apply OEM variables.

We enable this flag any time the device enters Fastboot mode or the Recovery
Console, since those are the two environments where new boot images may be
written. This isn't 100% accurate since it's possible the fastboot or
recovery session may not touch the boot image, but we don't care that much.

We also ensure when doing 'fastboot boot' that the temporary boot image's
oemvars are posted, but on the next normal boot they are restored.

Issue: GMINL-5555
Change-Id: If3e95580e3223399b73a865ddfc60beaecac3098
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>