use Life Cycle Protocol to detect the provisioning mode
If the OEMLock variable is undefined, Kernelflinger uses the Life
Cycle protocol to detect whether or not the device is in provisioning
mode. If this protocol is supported and the life cycle state is
EndUser, it sets provisioning_mode to FALSE, TRUE otherwise.

If the protocol is not supported or if we are on a non user build,
Kernelflinger assumes we are in provisioning mode.

Change-Id: I9dbc17d976e681f056754ae8ef33e799af86843e
Tracked-On: OAM-38333
Signed-off-by: Leo Sartre <leo.sartre@intel.com>
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: 542409