Add a parameter in oem command get-hashes to choose hash algorithm
The fastboot oem command get-hashes is currently using SHA1 algorithm
to compute image hashes, which is quite long.
The command is now taking an extra parameter to define the digest
algorithm to use. For now only SHA1 and MD5 are supported.
The time gain between SHA1 and MD5 is around 50%.

If no argument, the command will keep using SHA1 for backward
compatibility.

Change-Id: I7a64bbe3ad2475af9a3866ba00df8b81966c12d0
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-11592
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>