Persistent Kernelflinger error logs enabled for ENG & USERDEBUG builds
Users who don't have a debug board may still need to access the logs
in case of a KF crash so that they can report a bug with attached logs.
This patch enables them to do so by making the EFI-var stored logs
available even when not being in provisioning mode.
To get the logs, use this command :
fastboot oem get-provisioning-logs

Change-Id: I23a6c95cfb5f624f20733b161d94df6862561fcb
Tracked-On: GMINL-12234
Signed-off-by: Cyril Marpaud <cyrilx.marpaud@intel.com>