crashmode: provide a minimal adb implementation
In userdebug and eng build, and if the BIOS supports the USB device
mode, the "watchdog crash screen" now provides a minimal
implementation of the adb protocol.  Its implements the adb reboot
[TARGET] and pull commands as follow:

  - part:PART_NAME[:START[:LENGTH]]
  - acpi:TABLE_NAME
  - efivar:VAR_NAME[:GUID]
  - ram:[:START[:LENGTH]]

Where START and LENGTH are hexadecimal strings.  The ram command
output file is an Android sparse file.  Use the simg2img AOSP command
to get a flat binary.

The device is enumerated as "bootloader" (adb devices).

The adb protocol implementation lies in the libadb directory where
only the necessary has been implemented.  To add the support of new
pull commands, a new reader can very easily be implemented in
libadb/reader.c.

Tracked-On: OAM-873
Change-Id: Ica189684731b5b364cb756e661cd8c37e9412129
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: 401678