fastboot: support authenticated action
The 'oem get-action-nonce' command generates a nonce string including
the device serial number and a 16 bytes generated random number.

The user generates an action-authorization which is a PKCS7 message
signed with the OAK private key and including all the required
certificates.  The signed string include the nonce plus the action ID
and a 16 bytes random number.

Currently, the only supported action is force-unlock (ID 0x0) which
basically override the FRP mechanism and bootloader policy and force
the device to unlock.  Nevertheless, the same process than with the
usual unlock will happen (userdata partition will be erased).

The 'flash action-authorization' looks for the OAK certificate in the
PKCS7 packet based on the SHA256 stored in the OAK variable.  If
found, it verifies the PKCS7 against this certificate and extracts the
signed string.

Change-Id: I1f7e434175e8219dfa416b81345b00389dc04ea1
Tracked-On: OAM-860
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: 396392
Reviewed-on: 401444