bootloader-policy: do not power off the device on non-USER builds
Engineers needs to be able to boot USERDEBUG/ENG builds even if the
bootloader policy is class A or NO_DEVICE_UNLOCK and secure boot has
been disabled or boot state is yellow, orange or red.

The access granted because of a non-USER build is logged (serial
console) and printed on screen.

Change-Id: I7e62df59224546bc61d9ec87c7db1ebe08dc44fb
Tracked-On: OAM-2192
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: 409592