Cleanup load_boot_image()
This patch removes the validation part of load_boot_image().
Now the image have to be validated by a separate call to
validate_bootimage().

This simplifies the code because reduces the number of parameters of the
load_boot_image() function, and we can decide to verify the loaded image
or not, depending on the boot state.

When device is unlocked, the bootimages are not verified anymore.

Change-Id: I3ee06bfd753fb93292f68b9708129055fef969c9
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-2968
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>
Reviewed-on: https://android.intel.com:443/412515