Add support for RSCI v2
Note specific to COHO (BYT-CR):
COHO's BIOS does not generate a proper RSCI table compliant with
any of the revisions used for Cherrytrail or Broxton for instance.
While "not_applicable" is a valid reset source defined by an enum
as 0 in RSCI v1 and v2, COHO's BIOS considers any reset source with
value 0 as a "reboot". With changes done in this patch to support
RSCI v2, a build instruction is used to ignore the "not_applicable"
bootreason, and keep it to "reboot".
This was the decided approach to workaround the fact that COHO's BIOS
code cannot be changed anymore.

Change-Id: Id61a511b043ebe0f0f07768d6672bfb0a5abe861
Tracked-On: OAM-3692
Signed-off-by: Florent Auger <florent.auger@intel.com>
Reviewed-on: 415279