root of trust bitstream sha256
The bootimage validation results in a boot-state (GREEN, YELLOW,
ORANGE or RED).  From a security perspective, we also needs a unique
verification status to bind the root of trust to.

The Root of Trust bitstream must reflects both the device state and
the bootimage verification status.

The device-state is on one byte (0x1 for locked and 0x0 for unlocked).
The RoT-bitstream is the device-state followed by the RSA public key
of the certificate that has been used to verify the bootimage.

The resulting hash of the bootimage verification is
SHA256(RoT-bitstream).

Change-Id: Idde63aba32e910f90a62179fc6e30825016aa4ce
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-5177
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: https://android.intel.com:443/414257