Select power off when shutdown requested and panic or watchdog occurred
If a panic or a watchdog reset occurs after the user has requested
to shutdown the device, the reset consecutive to this event must not
reboot the platform to Main OS.
- On User builds only, Kernelflinger will detect this case and request
a "power off" instead.
- On other build types (userdebug/eng), reboot to keep debug mechanisms
available.

During shutdown, if panic happens, reboot reason EFI variable is
appended with the panic type.
At reboot,
- userdebug/eng : 2nd keyword (if available) is used as reboot reason
(ie "kernel_panic")
- all build types : 1st keyword is used as reboot_reason (ie "shutdown")

Change-Id: If901a6e52d67e52627a15c976d526b394d7d6c9f
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-7767
Signed-off-by: Francois-Nicolas Muller <francois-nicolas.muller@intel.com>
Reviewed-on: https://android.intel.com:443/433433