libfastboot: fill the first 4096 bytes with zeros on erase
If the Android fs_mgr fails mounting a partition, it tries to detect
if the partition has been wiped out to determine if it has to format
it.  fs_mgr considers that the partition has been wiped out if the
first 4096 bytes are filled up with all 0 or all 1.
storage_erase_blocks() uses hardware support to erase the blocks which
does not garantee that content will be all 0 or all 1.  It also can be
indeterminate data.

Tracked-On: OAM-7883
Change-Id: Icf46dfac7ebdcbbfbdd54eb4e18080922fee3faf
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: 433779