Make the binding root of trust when booting image with fastboot boot
"Fastboot boot" command load directly the image, without making the
root of trust binding.
In the normal boot flow, the root of trust binding is done even when
the device is in unlock state. In this case, the root of trust is
composed by the device_state only, since the verifier certificate
should not be checked in ORANGE boot flow.

cleanup: use die() function at the end of enter_fastboot_mode().

Change-Id: Ie850e9faa5d5c0c74fd9b028168827c5e91368f6
Tracked-On: OAM-10045
Signed-off-by: Leo Sartre <leox.sartre@intel.com>
Reviewed-on: 445190