Fix oem setvar command
When setting a variable using fastboot oem setvar and then setting the
same variable using fastboot flash oemvars, the last command fails with
the error "Invalid Parameter"

This is because the two ways doesn't use the same default value for the
variable attribute.

This patch align the two commands default attribute.

Change-Id: I2d76510ea96543a279c055893aefe2df2177f367
Tracked-On: OAM-14367
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>
Reviewed-on: 456359