Use silentlake vmm to bind RoT
Using hypercall to communicate the structure containing the device
state and the hash of the public key used to verify the boot
image. Those information are used by silentlake vmm as root of trust,
it will tie all cryptographic keys that it delivers with these
information.

The version field of the sctruct sl_gvb_data is versionning the data
structure itself and is used internally by the vmm. Starting with
VERSION 0, we will increment this number if the data structure got
changed.

The hypercall SL_CMD_HSEC_SET_GVB_INFO can be made only once, any
other attempt to call it without resetting the vmm will return an
error.

Change-Id: Ie56b23eb12c168b8b10aa478620897630acf0a38
Tracked-On: OAM-24289
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Signed-off-by: Leo Sartre <leox.sartre@intel.com>
Reviewed-on: 481313