Set OsSecureBoot variable at boot time
When boot state is green, kernelflinger will enable os secureboot,
otherwise it will disable it.
The os secureboot state will be read at ExitBootServices() call to
restrain or not the PMIC registers access.

The feature will be active only if KERNELFLINGER_OS_SECURE_BOOT is set
to true in the makefile.

Change-Id: If5cfe18aa74b42f319edbdb1b0069e0ebd6cb45d
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-10143
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>
Reviewed-on: https://android.intel.com:443/469728