crashmode: TCP support
This patch provides a new libefitcp to support TCP/IPv4 transport.
The libefitcp functions are asynchronous as the libefiusb functions
are : libefitcp also relies on RX and TX callback.

The libefitcp "start" function let the BIOS obtain the device IP
address from the network (DHCP/BOOTP).

A major difference between those two libraries is that the libefitcp
will wait until the requested data length as be read from the network
before calling the RX callback.

This patch makes crashmode (adb) uses this new TCP transport
implementation as a fallback if the USB transport is not supported.

Tracked-On: OAM-20874
Change-Id: I234a73fae1bcfb2d9855fce60f95064e009e365f
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: 472346