Remove TXE support
TXE is not the POR option for binding the root of trust.
This patch removes the support of txe and prepares the introduction of
the new implementation based on SilentLake.
The hash computation of the pub key of the certificate used to verify
the boot image was kept to display it in case of yellow boot scenario.

Change-Id: Idc61981624ea9f23a469c56989875e872e09e09c
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-24289
Signed-off-by: Leo Sartre <leox.sartre@intel.com>
Reviewed-on: https://android.intel.com:443/481312