fastboot: fix flash_fill() allocation failure
flash_fill() systematically allocates a buffer of SIZE bytes.  If size
is too large, the allocation fails.  This patch makes this function
allocates a reasonable size buffer and flash it as many times as
necessary.

This patch also prevents flash_write() to increase the cur_offset in
case of write failure.

Tracked-On: OAM-30713
Change-Id: Ie81008039c22e0cc58f22ecbbfc58f64d0d311c4
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: 501272