dm-verity in rootfs context
In rootfs context, the verity mode must be supplied by the bootloader
to the kernel.  This patch creates a new verity_corrupted field in the
slot metadata structure to remind that this slot has reported a
dm-verity corruption and must be started in EIO verity mode.  If this
field value is one, it prepends "androidboot.veritymode=eio" to the
kernel command line.

If a fastboot flash command related to a dm-verity partition (system,
vendor or oem) is flashed, the verity_corrupted is reset to 0.

Change-Id: Ib5ed67ea14b446df01d93b01c3428e340ac8d448
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-33775
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: https://android.intel.com:443/509316