Do not reset output/input console if there is no screen update
The output/input console is systematically resetted when booting
kernelflinger; meaning even if there is no vendor splash screen
enabled, the screen will be resetted to the default background instead
of keeping the firmware splash screen if there is one.

The output/input console is reseted in the ux_init_screen function
when there is an actual screen update, and only the first time.
If there is no screen update during kernelflinger boot, the screen
will stay as it was configured by firmware.

Change-Id: I8890823be751fb5df1df093e97adc0100425c4a2
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-33297
Reviewed-on: https://android.intel.com:443/514105