Update RED state to prevent from booting
When booting in RED state, the device should not boot anymore or power
off. It should enter bootloader recover mode instead.
Update the UX warning to specify the device won't boot.
Remove the 5s first timeout, and the message to warn about power off
or boot after a giving time.
The warning in that case will be displayed for 30s no matter what and
the volume up button won't have any effect.

Change-Id: I28afa8ff8afd2c8953f8fd518a71cea83c354362
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-33727
Reviewed-on: https://android.intel.com:443/514389