Update GVB UX to allow interacting with power button
The power button will be used to pause/continue the boot/shutdown of
the device instead of using the volume up button, except for the
platform that don't support the power button.
Moreover the new UX don't let any choice to the user to change the
ongoing sequence like before. It just allow to display the warning for
more or less time.

Change-Id: Ideb29957ed400c18bdc3829643632234f0003651
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-33804
Reviewed-on: https://android.intel.com:443/515320