Add a command to erase all optional efi vars
The command is available only for userdebug and eng builds, and in
unlocked state.
It allows to delete all the optional efi vars so they'll be resetted
to their default values.

The variable names are mutualized in vars.h.
The patch also fixes a minor error in cmd_oem_set_boolean() function.

Change-Id: Ia20d748fa6a31e76a4ce95e8ba55eb7bc02fa862
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>
Tracked-On: OAM-32855
Reviewed-on: 515768