gpt: support a mix of "android_" partitions and and not prefixed
OneAndroid adds the "android_" prefix to the Android partition labels
for the android partitions. However, we also have to support
non-android partitions which are not prefixed with the "android_"
string.  To support both case at the same time,
gpt_find_partition(LABEL) looks for both the requested LABEL and
L"android_" LABEL strings.

When exposing the partition information outside of this module we have
to make sure that this prefix has been removed.

Change-Id: I9d4f2c56955097240e93b989b3de49a78838b61a
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-44889
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: https://android.intel.com:443/579267