android: do NOT set up the efi_info structure on not UEFI system
The setup_memory_map() function is called by both UEFI and non UEFI
handover functions.  On non UEFI system, the System Table is powered
by a emulation library that is lost once Kernelflinger handover to the
Linux Kernel.

The patch makes sure setup_memory_map() does not populate the
boot_params efi_info structure if it is running on a non UEFI system.

Change-Id: I843a44ffa7273775bb559b25a3f8dbaaa2bf0245
Tracked-On: OAM-59119
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: 620114