Enable ql-tipc side to receive messages > 4K
This change allows the non-secure side to read a response that
may be larger than 4K by adding an additional bit indicating the
end of a response. If a message command has the KEYMASTER_STOP_BIT
set, then the non-secure side knows that the response has been fully read.

patch referenced from:
https://android-review.googlesource.com/c/trusty/external/trusty/+/503464

Change-Id: Ie2a37f993485e1a6da9093ba8b4306b8591637ee
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-60870
Signed-off-by: Yan, Shaopu <shaopu.yan@intel.com>
Reviewed-on: https://android.intel.com:443/625720