[Osloader] reboot to recovery mode on need
in case recovery mode, evmm and trusty should not be launched even if
trusty enabled.
in case normal boot, evmm and trusty will be launched if trusty enabled.

if recovery bcb found during normal boot, osloader need to launch recovery
os instead of normal boot. in this case, need to reboot to recovery mode
so evmm would not be launched by bootloader.

Change-Id: Iace1750add435b3664bd9eaf810a3ad75ace7a9f
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-70052
Signed-off-by: zhouji3x <jianfengx.zhou@intel.com>
Reviewed-on: https://android.intel.com:443/648555