redefine MEMORY boot target usage and behavior
With vbmeta and slots, the boot target MEMORY lost its purposes.  The
default fastboot boot behavior tested on Google Android Devices show
that the kernel should be taken from the downloaded boot image but
that the rootfs parameters, slots information, shoud be retrieve from
the eMMC.

This patch redefines the MEMORY boot target.  MEMORY boot target means
to load and boot the image loaded in memory regardless of what is on
the persistent storage.

Change-Id: Ibeacb6fd28cabc8a7e986e9bd55ef03c52548775
Tracked-On: OAM-75350
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Signed-off-by: "Sun, XunouX" <xunoux.sun@intel.com>
Reviewed-on: 658270