set-storage support to set UFS LUN number
In default, LUN numbers of UFS user and factory are hardcoded as 0 & 3.
This patch allows the command "oem set-storage ufs@lun<user>:<factory>"
in fastboot/installer to set these LUN numbers, for example:
    installer.efi oem set-storage ufs@lun1:3
this command sets user's LUN to 1 and factroy's to 3.

Include a refine: move the LUN detection of Boot image into probe().

Change-Id: I81ddbcd7b721accc9154fcf07258ebce9b945a12
Tracked-On:OAM-76644
Signed-off-by: Tang, Haoyu <haoyu.tang@intel.com>
Reviewed-on: 656452
Reviewed-on: 664189