usb: fix automatic transport layer fallback
Commit 52965e559cf48302af68b54b6d55ed3e514e07fa broke the transport
layer fallback mecanism.

The usb_start should run EFI_UNSUPPORTED if the no USB protocol are
support (and not EFI_NOT_FOUND).  The transport_start() function
depend on this particular error to decide to fallback on the next
available protocol implementation which is currently TCP if the UEFI
BIOS support this protocol.

Change-Id: I118197c8da912abe79371c45b087d8d4fb13b756
Tracked-On: OAM-80257
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: 668916