Fix issue that fastboot get a wrong boot hash value
The root cause of this issue is that boot partition size was not matched
with the boot image size as before. The hash value generation will be
affected by their size, therefore, hash value matched failure.

This patch adjusts boot image size variable to make sure that boot image
size is matched with boot partition size.

Change-Id: I17faaf39090ef68e825df9c8bd1f785dd99876be
Tracked-On: OAM-83865
Signed-off-by: sunxunou <xunoux.sun@intel.com>
Reviewed-on: 534332